Backfill origin_language/country data
There's a fair bit of complex logic involved here, but it's pretty much guessing based on a combination of titles, originalLocale, and subtype.

For more information, read the "Rename OEL to Global Manga" RFD, which this is roughly based on, though this only implements the first step and tries a bit harder to handle non-Japanese content than is proposed there.