Enable SFW filter on certain feeds
We don't want to show NSFW content on media pages or on global feeds, as this seems to just create shitshows constantly. We already have an SFW filter, this just enables it automatically beyond the user's setting. In future, we may provide different degrees of this setting to allow users to opt-in to this content.