Remove Chewy config which breaks on Ruby 3.x
Chewy is likely using an incorrect outdated form of splat in their non-after_commit implementation of hooks. This breaks in Ruby 3.x, but was only being used in our test environment.

The configuration line in question is 7 years old, so I don’t remember the exact context, but suspect it may have had something to do with the fact that each test is wrapped in a transaction which is never actually committed. In older versions of Rails, this meant your after_commit callbacks never executed, but that is no longer a relevant issue, being solved a few years ago