Remove YAML aliases from chewy.yml
Newer versions of Psych (introduced in Ruby 3.x) require users to opt-in to a number of advanced YAML features, including aliases, for safety reasons. Aliases can create an infinitely recursive structure, presenting surface area for a a potential denial-of-service attack.

Rails has opted into these features for their own YAML, but Chewy is basically unmaintained and has *not* opted into these features (because the opt-in didn’t exist back then) and so we just remove the usage of those features for now.